+++ train_SR.py
@@ -2,6 +2,7 @@
 script to train on SR classification dataset
 """
 import argparse
+import os
 import random
 import shutil
 import time
@@ -10,10 +11,11 @@ from json import dumps
 import numpy as np
 import torch
 from torch.optim import Adam
-from torch_geometric.data import DataLoader
+from torch_geometric.loader import DataLoader
+from torch_geometric.nn import DataParallel
 
 import train_utils
-from data_utils import extract_multi_hop_neighbors, PyG_collate, post_transform
+from data_utils import extract_multi_hop_neighbors, post_transform
 from datasets.SRDataset import SRDataset
 from layers.input_encoder import EmbeddingEncoder
 from layers.layer_utils import make_gnn_layer
@@ -21,28 +23,38 @@ from models.GraphClassification import GraphClassification
 from models.model_utils import make_GNN
 
 
-def train(train_loader, model, optimizer, device):
+def train(train_loader, model, optimizer, device, parallel=False):
     model.train()
     total_loss = 0
     for data in train_loader:
-        data = data.to(device)
         optimizer.zero_grad()
+        if parallel:
+            num_graphs = len(data)
+            y = torch.cat([d.y for d in data]).to(device)
+        else:
+            num_graphs = data.num_graphs
+            data = data.to(device)
+            y = data.y
         out = model(data).squeeze()
-        loss = torch.nn.CrossEntropyLoss()(out, data.y)
+        loss = torch.nn.CrossEntropyLoss()(out, y)
         loss.backward()
-        total_loss += loss.item() * data.num_graphs
+        total_loss += loss.item() * num_graphs
         optimizer.step()
     return total_loss / len(train_loader.dataset)
 
 
 @torch.no_grad()
-def test(loader, model, device):
+def test(loader, model, device, parallel=False):
     model.train()  # eliminate the effect of BN
     y_preds, y_trues = [], []
     for data in loader:
-        data = data.to(device)
+        if parallel:
+            y = torch.cat([d.y for d in data]).to(device)
+        else:
+            data = data.to(device)
+            y = data.y
         y_preds.append(torch.argmax(model(data), dim=-1))
-        y_trues.append(data.y)
+        y_trues.append(y)
     y_preds = torch.cat(y_preds, -1)
     y_trues = torch.cat(y_trues, -1)
     return (y_preds == y_trues).float().mean()
@@ -74,8 +86,8 @@ def get_model(args):
                                 output_size=args.output_size)
 
     model.reset_parameters()
-    # if args.parallel:
-    #    model = DataParallel(model, args.gpu_ids)
+    if args.parallel:
+        model = DataParallel(model, args.gpu_ids)
 
     return model
 
@@ -89,14 +101,15 @@ def main():
                         help='Probability of zeroing an activation in dropout layers.')
     parser.add_argument('--batch_size', type=int, default=64, help='Batch size per GPU. Scales automatically when \
                             multiple GPUs are available.')
+    parser.add_argument("--parallel", action="store_true",
+                        help="If true, use DataParallel for multi-gpu training")
     parser.add_argument('--num_workers', type=int, default=0, help='number of worker.')
     parser.add_argument('--load_path', type=str, default=None, help='Path to load as a model checkpoint.')
     parser.add_argument('--lr', type=float, default=0.001, help='Learning rate.')
-    parser.add_argument('--l2_wd', type=float, default=3e-6, help='L2 weight decay.')
+    parser.add_argument('--l2_wd', type=float, default=0., help='L2 weight decay.')
     parser.add_argument("--kernel", type=str, default="spd", choices=("gd", "spd"),
                         help="the kernel used for K-hop computation")
     parser.add_argument('--num_epochs', type=int, default=200, help='Number of epochs.')
-    parser.add_argument('--max_grad_norm', type=float, default=5.0, help='Maximum gradient norm for gradient clipping.')
     parser.add_argument("--hidden_size", type=int, default=48, help="hidden size of the model")
     parser.add_argument("--model_name", type=str, default="KPGIN",
                         choices=("KPGCN", "KPGIN", "KPGraphSAGE", "KPGINPlus"), help="Base GNN model")
@@ -155,10 +168,12 @@ def main():
     args.save_dir = train_utils.get_save_dir(args.save_dir, args.name, type=args.dataset_name)
     log = train_utils.get_logger(args.save_dir, args.name)
     device, args.gpu_ids = train_utils.get_available_devices()
-    if len(args.gpu_ids) > 1:
-        args.parallel = True
-    else:
-        args.parallel = False
+
+    log.info('multi-gpu training is prohibited due to the size '
+             'of SR25 dataset and unstable training performace')
+    args.parallel = False
+    loader = DataLoader
+
     args.batch_size *= max(1, len(args.gpu_ids))
 
     # Set random seed
@@ -168,7 +183,7 @@ def main():
     torch.manual_seed(args.seed)
     torch.cuda.manual_seed_all(args.seed)
 
-    def pre_transform(g):
+    def multihop_transform(g):
         return extract_multi_hop_neighbors(g, args.K, args.max_pe_num, args.max_hop_num, args.max_edge_type,
                                            args.max_edge_count,
                                            args.max_distance_count, args.kernel)
@@ -176,8 +191,9 @@ def main():
     transform = post_transform(args.wo_path_encoding, args.wo_edge_feature)
 
     path = "data/" + args.dataset_name
-    shutil.rmtree(path + '/processed')
-    dataset = SRDataset(path, pre_transform=pre_transform, transform=transform)
+    if os.path.exists(path + '/processed'):
+        shutil.rmtree(path + '/processed')
+    dataset = SRDataset(path, pre_transform=multihop_transform, transform=transform)
     dataset.data.x = dataset.data.x.long()
     dataset.data.y = torch.arange(len(dataset.data.y)).long()  # each graph is a unique class
     train_dataset = dataset
@@ -185,12 +201,9 @@ def main():
     test_dataset = dataset
 
     # 2. create loader
-    train_loader = DataLoader(train_dataset, args.batch_size, shuffle=True, num_workers=args.num_workers,
-                              collate_fn=PyG_collate)
-    val_loader = DataLoader(val_dataset, args.batch_size, shuffle=False, num_workers=args.num_workers,
-                            collate_fn=PyG_collate)
-    test_loader = DataLoader(test_dataset, args.batch_size, shuffle=False, num_workers=args.num_workers,
-                             collate_fn=PyG_collate)
+    train_loader = loader(train_dataset, args.batch_size, shuffle=True, num_workers=args.num_workers)
+    val_loader = loader(val_dataset, args.batch_size, shuffle=False, num_workers=args.num_workers)
+    test_loader = loader(test_dataset, args.batch_size, shuffle=False, num_workers=args.num_workers)
 
     # additional parameter for SR dataset and training
     args.input_size = 2
@@ -207,14 +220,14 @@ def main():
 
     optimizer = Adam(model.parameters(), lr=args.lr)
     best_val_acc, best_test_acc = 0, 0
+    start_outer = time.time()
     for epoch in range(args.num_epochs):
         start = time.time()
-        train_loss = train(train_loader, model, optimizer, device=device)
+        train_loss = train(train_loader, model, optimizer, device=device, parallel=args.parallel)
         lr = optimizer.param_groups[0]['lr']
-        log.info(f"evaluate after epoch {epoch + 1}...")
         model.eval()
-        val_acc = test(val_loader, model, device=device)
-        test_acc = test(test_loader, model, device=device)
+        val_acc = test(val_loader, model, device=device, parallel=args.parallel)
+        test_acc = test(test_loader, model, device=device, parallel=args.parallel)
         if val_acc >= best_val_acc:
             best_val_acc = val_acc
         if test_acc >= best_test_acc:
@@ -223,8 +236,13 @@ def main():
         time_per_epoch = time.time() - start
 
         log.info('Epoch: {:03d}, LR: {:7f}, Train Loss: {:.7f}, '
-                 'Val Acc: {:.7f}, Test Acc: {:.7f}, Best Val Acc: {:.7f}, Best Test Acc: {:.7f}, time: {:.4f}'.format(
+                 'Val Acc: {:.7f}, Test Acc: {:.7f}, Best Val Acc: {:.7f}, Best Test Acc: {:.7f}, Seconds: {:.4f}'.format(
             epoch + 1, lr, train_loss, val_acc, test_acc, best_val_acc, best_test_acc, time_per_epoch))
+        torch.cuda.empty_cache()  # empty test part memory cost
+
+    time_average_epoch = time.time() - start_outer
+    log.info(
+        f'Best val: {best_val_acc}, Best test: {best_test_acc}, Seconds/epoch: {time_average_epoch / (epoch+1)}')
 
 
 if __name__ == "__main__":
